Variable de indice para lazo de repeticion cambiado a uint8_t para ser consistente con el tipo de dato que se utiliza para totalLeds y totalSwitches

+++ bsp/edu_ciaa_nxp/src/board.c
@@ -76,7 +76,7 @@ static const int8_t totalSwitches = sizeof(switchMap) / sizeof(switchMap[0]);
 /*==================[external functions definition]==========================*/
 extern void board_init(void)
 {
-   int32_t i;
+   int8_t i;
 
    for (i = 0 ; i < totalLeds ; i++)
    {
